handler: Implement a custom std.json.Value parser
std.json.parseFromValueLeaky has to accept lots of input that should be
considered invalid as the implementation can't know what kind of input
the application expects from the struct it is given. The custom
implementation allows better error reporting and lets us handle all
Varlink value types in the future.