handler: Implement proper JSON serialization
The jsonize method can be used to serialize a zig-varlink style struct
to std.json.Value. This is useful for both implementing
serializeResponse and returning arbitrary objects to Varlink clients.

The downside of this is that the method requires an allocator, so the
handler's arena allocator is passed through the entire handler stack to
be able to give it to serializeResponse.