server: Track connection state
This makes the interface cleaner by keeping track of the correct
response type and reducing the number of passed arguments.

Tracking connection state is the start of turning the server isto a
nicer and more usable interface. For instance, it makes it possible to
add more safety checks and helpers for continuation replies and oneway
by keeping track of whether they are set or not. Additionally, it makes
it possible to get rid of using a writer directly and allows letting
library users manage their message format more easily.